Move the keepTab service to the wp-panels directory
