+++ frontend/app/components/wp-buttons/wp-details-view-button/wp-details-view-button.directive.ts
@@ -27,7 +27,7 @@
 // ++
 
 import {WorkPackageNavigationButtonController, wpButtonDirective} from '../wp-buttons.module';
-import {KeepTabService} from '../keep-tab/keep-tab.service';
+import {KeepTabService} from '../../wp-panels/keep-tab/keep-tab.service';
 
 export class WorkPackageDetailsViewButtonController extends WorkPackageNavigationButtonController {
   public projectIdentifier:number;
