+++ frontend/app/components/wp-panels/keep-tab/keep-tab.service.ts
@@ -59,5 +59,5 @@ export class KeepTabService {
 }
 
 angular
-  .module('openproject.wpButtons')
+  .module('openproject.workPackages.services')
   .service('keepTab', KeepTabService);
